Use Fog DNSimple API Auth
The Fog gem now accepts an API token for creating DNSimple records.
The API Token method is preferred so you can use 2FA auth as well as
not have your password log to console.

The Fog gem currently accepts the token as an additional param. If it
detects the password param it will fall back to username/password auth.