Update the README with ALL THE THINGS
We removed the attributes section because it turns out we don't actually
use them anywhere in the LWRP. In the future those will be removed
anyway because exposed credentials in the attributes is no bueno for
security.

This also documents the upcoming deprecations along with notes about API
usage and best practices for security with chef vault examples.