Rip out webmock entirely
I don't believe the edges of our library include the HTTP side of
things. I think if we're going to control for anything it should be at
the dnsimple api gem. How it fetches and grabs information isn't what we
as a consumer care about. We care about how it returns data to us and
that is it's contract with us.